Propsの渡し方、emitの実装方法
